+++ plugins/sql-gemini.php
@@ -9,7 +9,7 @@
 * @license https://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
 * @license https://www.gnu.org/licenses/gpl-2.0.html GNU General Public License, version 2 (one or other)
 */
-class AdminerSqlGemini {
+class AdminerSqlGemini extends Adminer\Plugin {
 	private $apiKey;
 	private $model;
 
@@ -92,11 +92,7 @@ geminiText.onkeydown = event => {
 <?php
 	}
 
-	private function lang($idf, $number = null) {
-		return Adminer\lang_format(Adminer\idx(self::$translations[Adminer\LANG], $idf) ?: $idf, $number);
-	}
-
-	private static $translations = array(
+	protected static $translations = array(
 		'cs' => array(
 			'Ask Gemini' => 'Zeptat se Gemini',
 			'Just a sec...' => 'Chviličku...',